+++ Entities/PEmitter.cpp
@@ -157,7 +157,7 @@ namespace RTE {
 			reader >> burstSize;
 			// Go through all emissions and set the rate so that it emulates the way it used to work, for mod backwards compatibility
 			for (list<Emission>::iterator eItr = m_EmissionList.begin(); eItr != m_EmissionList.end(); ++eItr)
-				(*eItr).m_BurstSize = ceilf((float)burstSize / (float)m_EmissionList.size());
+				(*eItr).m_BurstSize = std::ceil((float)burstSize / (float)m_EmissionList.size());
 		}
 		else if (propName == "BurstScale")
 			reader >> m_BurstScale;
